стабильная версия, перед внедрением зависимостей и рефакторингом

+++ app/src/main/java/com/stanislavkorneev/korneevapp/domain/entities/LoanConditions.kt
@@ -1,7 +1,7 @@
 package com.stanislavkorneev.korneevapp.domain.entities
 
 data class LoanConditions(
-    val maxAmount: Double,
+    val maxAmount: Int,
     val percent: Double,
     val period: Int,
 )
